% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Look up definition of term.}
\usage{
define(key, lang = "en", glossary = NULL, show_lang = FALSE)
}
\arguments{
\item{key}{lookup key}

\item{lang}{language for the definition}

\item{glossary}{a glossary object}

\item{show_lang}{should the ISO639-1 2-letter code be displayed with the
definition?}
}
\value{
string containing definition
}
\description{
Look up definition of term.
}
