% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_terms.R
\name{find_terms}
\alias{find_terms}
\title{Find terms in lessons}
\usage{
find_terms(files = ".", terms = NULL, verbose = TRUE)
}
\arguments{
\item{files}{The vector of specific files you want scanned or the
top-level folder to start searching for lessons}

\item{terms}{The vector of the terms you want to find}

\item{verbose}{If to print each term when found or not, TRUE by default}
}
\value{
list of which terms were found and in which files
}
\description{
Find terms in lessons
}
\details{
This functions helps the user find terms in different lessons, irrespective
of folder management strategy or organization hierarchy. If the term is found,
a list entry will contain the term as a key and the files as values.
}
