% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glosario.R
\docType{package}
\name{glosario-package}
\alias{glosario-package}
\title{A multilingual glossary.}
\description{
\code{glosario} provides the infrastructure to create and read multilingual
glossaries. By default, \code{glosario} providesaccess to a community-curated
glossary hosted by \href{https://github.com/carpentries/glosario}{The Carpentries}.
}
