% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_glosario_terms.R
\name{list_glosario_terms}
\alias{list_glosario_terms}
\title{List all the glosario terms in an RMd}
\usage{
list_glosario_terms(title = NULL, type = "definitions")
}
\arguments{
\item{title}{The table title}

\item{type}{One of 'definitions' or 'reqs'}
}
\description{
List all the glosario terms in an RMd
}
