% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Look up definition of term.}
\usage{
define(key, lang = "en", glossary, show_lang = FALSE)
}
\arguments{
\item{key}{lookup key}

\item{lang}{language for the definition}

\item{glossary}{a glossary object}
}
\value{
string containing definition
}
\description{
Look up definition of term.
}
