% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdef.R
\name{gdef}
\alias{gdef}
\title{Generate anchor tag for a given term in HTML}
\usage{
gdef(term, text)
}
\arguments{
\item{term}{Look-up key}

\item{text}{Text to display}
}
\value{
string containing link
}
\description{
Generate anchor tag for a given term in HTML
}
\note{
If you set up a language key in the glosario section of the YAML header,
the function will read this key and link to the correct language. If no
language key is provided the function will default to English.
}
