% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_glossary}
\alias{get_glossary}
\title{Get a glossary}
\usage{
get_glossary(
  url = "https://raw.githubusercontent.com/gvwilson/glossary/master/glossary.yml",
  cache = tempdir()
)
}
\arguments{
\item{url}{the path for the glossary}
}
\value{
a list
}
\description{
Retrive a YAML-formatted glossary from the web or locally.
}
\details{
The specification of the file is available from https://github.com/gvwilson/glossary/#readme
}
